refactor (chat box): add ui_kit Element for textarea
This allows the textarea, used to input chat messages, to be reused.
